\name{miRNA_target_package-package}
\alias{miRNA_target_package-package}
\alias{miRNA_target_package}
\docType{package}
\title{
\packageTitle{miRNA_target_package}
}
\description{
\packageDescription{miRNA_target_package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{miRNA_target_package}
\packageIndices{miRNA_target_package}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{miRNA_target_package}

Maintainer: \packageMaintainer{miRNA_target_package}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
